Change Id and Position type from Int to Long. Change license file header. Other fixes.

+++ VanKornoDB/src/main/java/com/vankorno/vankornodb/dbManagement/data/BaseEntity.kt
@@ -5,7 +5,7 @@ import com.vankorno.vankornodb.api.OldEntity
 
 
 interface BaseEntity {
-    val id: Int
+    val id: Long
 }
 
 interface NormalEntity : BaseEntity
@@ -17,11 +17,11 @@ interface NormalEntity : BaseEntity
 // Mainly for tests:
 
 abstract class CurrEntityWithId : CurrEntity {
-    override val id: Int = -1
+    override val id: Long = -1L
 }
 
 
 abstract class OldEntityWithId : OldEntity {
-    override val id: Int = -1
+    override val id: Long = -1L
 }
